+++ app/src/main/java/com/yakushev/spring/domain/model/ApplePointModel.kt
@@ -1,13 +1,13 @@
 package com.yakushev.spring.domain.model
 
 data class ApplePointModel(
-    val x: Int,
-    val y: Int,
+    val x: Float,
+    val y: Float,
 ) {
     companion object {
         val empty = ApplePointModel(
-            x = 0,
-            y = 0,
+            x = 0f,
+            y = 0f,
         )
     }
 }
